Added tabs for past and upcoming sessions, changed "slots" to "sessions" in UI, changed UI to make it consistent (buttons, text color in buttons) and renamed components
